requirements.txt
defusedxml==0.4.1
dj-database-url==0.4.1
django-extensions==1.6.7
django-filter==0.9.2
django-test-without-migrations==0.4
Django==1.9.6
djangorestframework==3.3.3
gunicorn==19.5.0
markdown==2.1.0
oauthlib==1.1.1
psycopg2==2.6.1
PyJWT==1.4.0
python-decouple==3.0
python-social-auth==0.2.14
python3-openid==3.0.10
pytz==2016.4
requests-oauthlib==0.6.1
requests==2.10.0
six==1.10.0
whitenoise==3.1
